fix/devide chat function
